mode logs APIs to seperate file
