+++ src/index.js
@@ -10,6 +10,7 @@ import shipping from './api/shipping';
 import stores from './api/stores';
 import subscriptions from './api/subscriptions';
 import taxes from './api/taxes';
+import logs from './api/logs';
 
 KnawatMP._addToPrototype(categories);
 KnawatMP._addToPrototype(coupons);
@@ -22,5 +23,6 @@ KnawatMP._addToPrototype(shipping);
 KnawatMP._addToPrototype(stores);
 KnawatMP._addToPrototype(subscriptions);
 KnawatMP._addToPrototype(taxes);
+KnawatMP._addToPrototype(logs);
 
 module.exports = KnawatMP;
